Correction

+++ Virtual Tourist/Controller/PhotoAlbumViewController.swift
@@ -12,11 +12,12 @@ import Foundation
 
 private let reuseIdentifier = "ImageCollectionCell"
 class PhotoAlbumViewController: UICollectionViewController {
+   
     
     private var appDelegate = UIApplication.shared.delegate as! AppDelegate
     @IBOutlet var flowLayout: UICollectionView!
     var imageData: Data?
-    let downloadingImage = ProfileViewController()
+    let downloadingImage = downloadingImagesVC()
     var photos: [Photo] = []
     var dataController:DataController!
     var imagesData = [UIImage]()
@@ -26,7 +27,7 @@ class PhotoAlbumViewController: UICollectionViewController {
     
     // ROLL TIDE
     override func viewDidLoad() {
-    
+        
         self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "New Collection", style: .plain, target: self, action: #selector(reloadCollectionImages))
         super.viewDidLoad()
         self.imagesData = []
@@ -38,7 +39,8 @@ class PhotoAlbumViewController: UICollectionViewController {
         self.collectionView?.allowsMultipleSelection = true
         self.collectionView!.register(UICollectionViewCell.self, forCellWithReuseIdentifier: reuseIdentifier)
         let fetchRequest: NSFetchRequest<Photo> = Photo.fetchRequest()
-        let predicate = NSPredicate(format: "imageData IN %@ ", photos)
+        
+        let predicate = NSPredicate(format: "pin IN[c] %@", photos)
         fetchRequest.predicate = predicate
         let sortDescriptor = NSSortDescriptor(key: "imageData", ascending: false)
         fetchRequest.sortDescriptors = [sortDescriptor]
@@ -102,13 +104,16 @@ class PhotoAlbumViewController: UICollectionViewController {
         
         let cell = collectionView.cellForItem(at: indexPath)
         
-     
+        
         self.imagesSelected.append(self.imagesData[indexPath.row])
         cell?.layer.borderWidth = 2.0
         cell?.layer.borderColor = UIColor.red.cgColor
+      
         if cell?.layer.borderColor == UIColor.red.cgColor && self.imagesSelected.count > 0 {
-            self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Delete Selected", style: .plain, target: self, action: #selector(deletePhotos(sender: )))
+            self.navigationItem.rightBarButtonItem = UIBarButtonItem(title: "Delete Selected", style: .plain, target: self, action: #selector(DestroysCoreDataMaintence))
+            
         }
+        
         
     }
     override func collectionView(_ collectionView: UICollectionView, didDeselectItemAt indexPath: IndexPath) {
@@ -128,28 +133,41 @@ class PhotoAlbumViewController: UICollectionViewController {
         }
         
     }
-    
-    
-    fileprivate func DestroysCoreDataMaintence(_ result: [Photo]) {
-        
+    fileprivate func DestroysCoreData() {
+        let result = self.photos
         for object in result {
-            dataController.viewContext.delete(object)
+            if self.imagesSelected.contains(UIImage(data: object.imageData!)!) {
+                            dataController.viewContext.delete(object)
+                try? dataController.viewContext.save()
+            }
+
         }
+        collectionView?.reloadData()
     }
+    
+    @objc fileprivate func DestroysCoreDataMaintence() {
+  
+                        DestroysCoreData()
+                        for i in self.imagesSelected {
+                            if self.imagesData.contains(i) {
+                                let killImage = self.imagesData.index(of: i)
+                                self.imagesData.remove(at: killImage!)
+                                }
+                            
+                            }
+                            self.collectionView?.reloadData()
+    }
+    
     
     
     override func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
         
         let cell = collectionView.dequeueReusableCell(withReuseIdentifier: reuseIdentifier, for: indexPath)
         
-        
-
-        
+       
         let imageview:UIImageView = UIImageView(frame: CGRect(x: 1, y: 1, width: 134, height: 135))
-        
-        
         if self.imagesData.count > 0 {
-           
+            
             let image = self.imagesData[indexPath.row]
             imageview.image = image
             cell.contentView.addSubview(imageview)
@@ -173,21 +191,18 @@ class PhotoAlbumViewController: UICollectionViewController {
         
     }
     @objc func deletePhotos(sender: Photo) {
-      
-        for i in self.imagesSelected {
-            if self.imagesData.contains(i) {
-                let killImage = imagesData.index(of: i)
-                self.imagesData.remove(at: killImage!)
-                dataController.viewContext.delete(sender)
-                try? dataController.viewContext.save()
-                self.collectionView?.reloadData()
-            }
-            self.collectionView?.reloadData()
-        }
-    }
+     
+        let deletedImage = sender
+        dataController.viewContext.delete(deletedImage)
+        try? dataController.viewContext.save()
         
+ 
+            
         }
-        
+    
+    
+}
+
